unit tests for finders locator

+++ src/core/utils.ts
@@ -5,6 +5,6 @@
  * @param {Uint8ClampedArray} arr - Input array.
  * @returns - The sum of the array.
  */
-export function sumArray(arr: Uint8ClampedArray): number {
+export function sumArray(arr: Int32Array): number {
     return arr.reduce((sum, x) => sum + x);
 }
